attrs.FieldDef now adheres to sql.Scanner and driver.Valuer
this allows for automatic conversions in the database, I.E. attrs.FieldDef.Value() will return a primary key if the value is of type attrs.Definer
scan can then be called on said primary key value, which in turn will retrieve the contenttype definition from the type of the field, allowing for fetching the full object when scanning
todo: somehow make this lazy? seems like it would be very intensive on the database.